backend/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.7
pyyaml>=6.0
jsonschema>=4.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
